MPS-9336 [build:11.SNAPSHOT] User's action execute method failed. Action:DeleteModels

+++ workbench/mps-workbench/source/jetbrains/mps/workbench/actions/model/DeleteModelHelper.java
@@ -69,7 +69,11 @@ public class DeleteModelHelper {
   }
 
   public static void deleteGeneratedFiles(SModelDescriptor modelDescriptor) {
-    IFile moduleOutput = FileSystem.getInstance().getFileByPath(modelDescriptor.getModule().getOutputFor(modelDescriptor));
+    String moduleOutputPath = modelDescriptor.getModule().getOutputFor(modelDescriptor);
+    if (moduleOutputPath == null) {
+      return;
+    }
+    IFile moduleOutput = FileSystem.getInstance().getFileByPath(moduleOutputPath);
     FileGenerationUtil.getDefaultOutputDir(modelDescriptor, moduleOutput).delete();
     FileGenerationUtil.getDefaultOutputDir(modelDescriptor, FileGenerationUtil.getCachesDir(moduleOutput)).delete();
   }
